06_게시판배포/requirements.txt
asgiref==3.3.1
autopep8==1.5.6
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
django-environ==0.4.5